LPS-196642 Inline CommerceSearchUtil methods in CommerceSearchResource
