+++ modules/apps/commerce/commerce-frontend-impl/src/main/java/com/liferay/commerce/frontend/internal/search/CommerceSearchResource.java
@@ -10,8 +10,10 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 
 import com.liferay.account.constants.AccountConstants;
+import com.liferay.account.constants.AccountPortletKeys;
 import com.liferay.account.model.AccountEntry;
 import com.liferay.account.service.AccountGroupLocalService;
+import com.liferay.commerce.constants.CommercePortletKeys;
 import com.liferay.commerce.context.CommerceContext;
 import com.liferay.commerce.context.CommerceContextFactory;
 import com.liferay.commerce.frontend.internal.account.CommerceAccountResource;
@@ -21,11 +23,11 @@ import com.liferay.commerce.frontend.internal.account.model.Order;
 import com.liferay.commerce.frontend.internal.account.model.OrderList;
 import com.liferay.commerce.frontend.internal.order.CommerceOrderResource;
 import com.liferay.commerce.frontend.internal.search.model.SearchItemModel;
-import com.liferay.commerce.frontend.internal.search.util.CommerceSearchUtil;
 import com.liferay.commerce.model.CommerceOrder;
 import com.liferay.commerce.order.CommerceOrderHttpHelper;
 import com.liferay.commerce.product.catalog.CPCatalogEntry;
 import com.liferay.commerce.product.catalog.CPQuery;
+import com.liferay.commerce.product.constants.CPPortletKeys;
 import com.liferay.commerce.product.data.source.CPDataSourceResult;
 import com.liferay.commerce.product.model.CommerceChannel;
 import com.liferay.commerce.product.service.CommerceChannelLocalService;
@@ -38,6 +40,7 @@ import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.Layout;
+import com.liferay.portal.kernel.model.LayoutSet;
 import com.liferay.portal.kernel.portlet.PortletProvider;
 import com.liferay.portal.kernel.portlet.PortletProviderUtil;
 import com.liferay.portal.kernel.search.Field;
@@ -49,6 +52,7 @@ import com.liferay.portal.kernel.util.HashMapBuilder;
 import com.liferay.portal.kernel.util.HtmlParser;
 import com.liferay.portal.kernel.util.HtmlUtil;
 import com.liferay.portal.kernel.util.HttpComponentsUtil;
+import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
@@ -113,7 +117,7 @@ public class CommerceSearchResource {
 					_searchOrders(queryString, themeDisplay));
 			}
 
-			String url = _commerceSearchUtil.getSearchFriendlyURL(themeDisplay);
+			String url = _getSearchFriendlyURL(themeDisplay);
 
 			if (Validator.isNotNull(url)) {
 				url = HttpComponentsUtil.addParameter(url, "q", queryString);
@@ -142,6 +146,42 @@ public class CommerceSearchResource {
 		).build();
 	}
 
+	private String _getAccountManagementFriendlyURL(ThemeDisplay themeDisplay)
+		throws PortalException {
+
+		LayoutSet layoutSet = themeDisplay.getLayoutSet();
+
+		Layout layout = _getAccountManagementLayout(
+			themeDisplay.getScopeGroupId(), layoutSet.isPrivateLayout());
+
+		if (layout == null) {
+			return null;
+		}
+
+		return _portal.getLayoutFriendlyURL(layout, themeDisplay);
+	}
+
+	private Layout _getAccountManagementLayout(
+			long groupId, boolean privateLayout)
+		throws PortalException {
+
+		Layout layout = _layoutLocalService.fetchLayoutByFriendlyURL(
+			groupId, privateLayout, "/accounts");
+
+		if (layout != null) {
+			return layout;
+		}
+
+		long plid = _portal.getPlidFromPortletId(
+			groupId, AccountPortletKeys.ACCOUNT_ENTRIES_MANAGEMENT);
+
+		if (plid > 0) {
+			layout = _layoutLocalService.fetchLayout(plid);
+		}
+
+		return layout;
+	}
+
 	private String _getAccountManagementPortletEditURL(
 			long accountId, ThemeDisplay themeDisplay)
 		throws PortalException {
@@ -159,6 +199,41 @@ public class CommerceSearchResource {
 		return editURL.toString();
 	}
 
+	private String _getCatalogFriendlyURL(ThemeDisplay themeDisplay)
+		throws PortalException {
+
+		LayoutSet layoutSet = themeDisplay.getLayoutSet();
+
+		Layout layout = _getCatalogLayout(
+			themeDisplay.getScopeGroupId(), layoutSet.isPrivateLayout());
+
+		if (layout == null) {
+			return null;
+		}
+
+		return _portal.getLayoutFriendlyURL(layout, themeDisplay);
+	}
+
+	private Layout _getCatalogLayout(long groupId, boolean privateLayout)
+		throws PortalException {
+
+		Layout layout = _layoutLocalService.fetchLayoutByFriendlyURL(
+			groupId, privateLayout, "/catalog");
+
+		if (layout != null) {
+			return layout;
+		}
+
+		long plid = _portal.getPlidFromPortletId(
+			groupId, CPPortletKeys.CP_SEARCH_RESULTS);
+
+		if (plid > 0) {
+			layout = _layoutLocalService.fetchLayout(plid);
+		}
+
+		return layout;
+	}
+
 	private CommerceOrder _getCommerceOrder(long commerceOrderId) {
 		try {
 			return _commerceOrderService.getCommerceOrder(commerceOrderId);
@@ -170,6 +245,66 @@ public class CommerceSearchResource {
 
 			return null;
 		}
+	}
+
+	private String _getOrdersFriendlyURL(ThemeDisplay themeDisplay)
+		throws PortalException {
+
+		LayoutSet layoutSet = themeDisplay.getLayoutSet();
+
+		Layout layout = _getOrdersLayout(
+			themeDisplay.getScopeGroupId(), layoutSet.isPrivateLayout());
+
+		if (layout == null) {
+			return null;
+		}
+
+		return _portal.getLayoutFriendlyURL(layout, themeDisplay);
+	}
+
+	private Layout _getOrdersLayout(long groupId, boolean privateLayout)
+		throws PortalException {
+
+		Layout layout = _layoutLocalService.fetchLayoutByFriendlyURL(
+			groupId, privateLayout, "/cart");
+
+		if (layout == null) {
+			layout = _layoutLocalService.fetchLayoutByFriendlyURL(
+				groupId, privateLayout, "/orders");
+		}
+
+		if (layout != null) {
+			return layout;
+		}
+
+		long plid = _portal.getPlidFromPortletId(
+			groupId, CommercePortletKeys.COMMERCE_CART_CONTENT);
+
+		if (plid <= 0) {
+			plid = _portal.getPlidFromPortletId(
+				groupId, CommercePortletKeys.COMMERCE_ORDER_CONTENT);
+		}
+
+		if (plid > 0) {
+			layout = _layoutLocalService.fetchLayout(plid);
+		}
+
+		return layout;
+	}
+
+	private String _getSearchFriendlyURL(ThemeDisplay themeDisplay)
+		throws PortalException {
+
+		LayoutSet layoutSet = themeDisplay.getLayoutSet();
+
+		Layout layout = _getSearchLayout(
+			themeDisplay.getScopeGroupId(), layoutSet.isPrivateLayout());
+
+		if (layout == null) {
+			return null;
+		}
+
+		return _portal.getLayoutFriendlyURL(layout, themeDisplay);
 	}
 
 	private SearchItemModel _getSearchItemModel(
@@ -196,6 +331,11 @@ public class CommerceSearchResource {
 				cpCatalogEntry.getCPDefinitionId(), themeDisplay));
 
 		return searchItemModel;
+	}
+
+	private Layout _getSearchLayout(long groupId, boolean privateLayout) {
+		return _layoutLocalService.fetchLayoutByFriendlyURL(
+			groupId, privateLayout, "/search");
 	}
 
 	private List<SearchItemModel> _searchAccounts(
@@ -235,8 +375,7 @@ public class CommerceSearchResource {
 			searchItemModels.add(searchItemModel);
 		}
 
-		String url = _commerceSearchUtil.getAccountManagementFriendlyURL(
-			themeDisplay);
+		String url = _getAccountManagementFriendlyURL(themeDisplay);
 
 		if (Validator.isNotNull(url)) {
 			url = HttpComponentsUtil.addParameter(url, "q", queryString);
@@ -286,7 +425,7 @@ public class CommerceSearchResource {
 			searchItemModels.add(searchItemModel);
 		}
 
-		String url = _commerceSearchUtil.getOrdersFriendlyURL(themeDisplay);
+		String url = _getOrdersFriendlyURL(themeDisplay);
 
 		if (Validator.isNotNull(url)) {
 			url = HttpComponentsUtil.addParameter(url, "q", queryString);
@@ -370,7 +509,7 @@ public class CommerceSearchResource {
 					accountEntryId, cpCatalogEntry, themeDisplay));
 		}
 
-		String url = _commerceSearchUtil.getCatalogFriendlyURL(themeDisplay);
+		String url = _getCatalogFriendlyURL(themeDisplay);
 
 		if (Validator.isNotNull(url)) {
 			url = HttpComponentsUtil.addParameter(url, "q", queryString);
@@ -421,9 +560,6 @@ public class CommerceSearchResource {
 	private CommerceOrderService _commerceOrderService;
 
 	@Reference
-	private CommerceSearchUtil _commerceSearchUtil;
-
-	@Reference
 	private CPDefinitionHelper _cpDefinitionHelper;
 
 	@Reference
@@ -434,5 +570,8 @@ public class CommerceSearchResource {
 
 	@Reference
 	private LayoutLocalService _layoutLocalService;
+
+	@Reference
+	private Portal _portal;
 
 }
