+++ modules/apps/commerce/source-formatter-suppressions.xml
@@ -16,7 +16,6 @@
 		<suppress checks="BNDExportsCheck" files="commerce-service/bnd\.bnd" />
 		<suppress checks="BNDExportsCheck" files="commerce-shipping-web/bnd\.bnd" />
 		<suppress checks="JavaConfigurationCategoryCheck" files=".*" />
-		<suppress checks="JavaHelperUtilCheck" files="commerce-frontend-impl/src/main/java/com/liferay/commerce/frontend/internal/search/util/CommerceSearchUtil\.java" />
 		<suppress checks="JavaHelperUtilCheck" files="commerce-product-asset-categories-web/src/main/java/com/liferay/commerce/product/asset/categories/web/internal/util/CPAssetCategoryWebPortletUtil\.java" />
 		<suppress checks="JavaUpgradeVersionCheck" files="commerce-product-type-virtual-service/src/main/java/com/liferay/commerce/product/type/virtual/internal/upgrade/registry/CommerceProductTypeVirtualServiceUpgradeStepRegistrator\.java" />
 		<suppress checks="JavaUpgradeVersionCheck" files="commerce-shipping-engine-fixed-service/src/main/java/com/liferay/commerce/shipping/engine/fixed/internal/upgrade/registry/CommerceShippingEngineFixedServiceUpgradeStepRegistrator\.java" />
